requirements_tpu.txt
absl-py==0.9.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
dill==0.3.1.1
future==0.18.2
googleapis-common-protos==1.51.0
idna==2.8
numpy==1.18.1
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
promise==2.3
protobuf==3.11.3
requests==2.22.0
six==1.14.0
tensorflow-datasets==2.0.0
tensorflow-metadata==0.21.1
termcolor==1.1.0
tqdm==4.42.1
urllib3==1.25.8
wrapt==1.11.2